Move code to appropriate files to match current filing format
