+++ galley/formatted_queries.py
@@ -1,10 +1,11 @@
 from typing import Dict, Optional, List
 
-from galley.queries import get_raw_recipes_data
+from galley.queries import get_raw_recipes_data, get_raw_menu_data
 from galley.pagination import paginate_results
 
 FOOD_PACKAGING = 'food pkg'
 STANDALONE = 'standalone'
+
 
 def ingredients_from_recipe_items(recipe_items: List[Dict]) -> Optional[List]:
     ingredients = []
@@ -35,9 +36,22 @@ def ingredients_from_recipe_items(recipe_items: List[Dict]) -> Optional[List]:
                 for ingredient in item_ingredients:
                     if ingredient not in ingredients:
                         ingredients.append(ingredient)
-
     return ingredients
 
+
+# Returns the subRecipeId if any 'standalone' item exists within recipeItems, else returns None
+# It is assumed that there is a max of ONE standalone item within the list of recipeItems, if any.
+def get_standalone(recipe_items):
+    for recipe_item in recipe_items:
+        preparations = recipe_item.get('preparations', [])
+        is_standalone = any(prep['name'] == STANDALONE for prep in preparations)
+
+        if is_standalone:
+            return recipe_item.get('subRecipeId')
+    return None
+
+
+# DATA TRANSFORMATION FUNCTIONS
 
 @paginate_results()
 def get_formatted_recipes_data(recipe_ids: List[str]) -> Optional[List[Dict]]:
@@ -68,5 +82,35 @@ def get_formatted_recipes_data(recipe_ids: List[str]) -> Optional[List[Dict]]:
         formatted_recipe['ingredients'] = ingredients_from_recipe_items(recipe_items=recipe_items)
 
         formatted_recipes.append(formatted_recipe)
-
     return formatted_recipes
+
+
+@paginate_results()
+def get_formatted_menu_data(names: list) -> Optional[List[Dict]]:
+    menus = get_raw_menu_data(names)
+    formatted_menus = []
+
+    if not menus:
+        return None
+
+    for menu in menus:
+        formatted_menu = ({
+            'name': menu.get('name'),
+            'id': menu.get('id'),
+            'date': menu.get('date'),
+            'location': menu['location'].get('name'),
+            'menuItems': []
+        }) # type: Dict
+
+        menu_items = menu.get('menuItems', [])
+        for menu_item in menu_items:
+            recipe_items = menu_item.get('recipe', {}).get('recipeItems', [])
+
+            formatted_menu['menuItems'].append({
+                'itemCode': next((cv.get('name') for cv in menu_item['categoryValues']), None),
+                'recipeId': menu_item.get('recipeId'),
+                'standaloneRecipeId': get_standalone(recipe_items)
+            })
+
+        formatted_menus.append(formatted_menu)
+    return formatted_menus
