+++ galley/pagination.py
@@ -25,9 +25,11 @@ def paginate_results(page_size=25):
                 return result
 
             for partition in range((len(input_list) // page_size) + 1):
-                chunk = input_list[partition * page_size: (partition + 1) * page_size]
-                result.extend(func(chunk)) if len(chunk) > 0 else None
-
+                try:
+                    chunk = input_list[partition * page_size: (partition + 1) * page_size]
+                    result.extend(func(chunk)) if len(chunk) > 0 else None
+                except Exception as e:
+                    logger.error(e)
             return result
         return wrapper
     return decorating_function
